+++ src/image/tilemap_entry.rs
@@ -110,7 +110,7 @@ impl TilemapEntry {
         (self.0 & 0x3FF)
             + (if self.1 { 1 } else { 0 } << 10)
             + (if self.2 { 1 } else { 0 } << 11)
-            + ((self.3 as usize & 0x3F) << 12) as usize
+            + ((self.3 as usize & 0x3F) << 12)
     }
 }
